% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{countries}
\alias{countries}
\title{A \code{data.frame} containing the names of countries around the world.}
\format{
\code{data.frame} containing 250 rows and 6 variables
\describe{
\item{iso2}{The 2 letter country code (ISO-3166-1)}
\item{iso3}{The 3 letter country code (ISO-3166-1)}
\item{isoNumerical}{The ISO-3166-1 number}
\item{title}{The title of each country in the world}
\item{gbifRegion}{The region/contient that pertains to gbif records}
\item{enumName}{The enumerated name used by gbif}
}
}
\usage{
countries
}
\description{
A \code{data.frame} containing the names of countries around the world.
}
\keyword{datasets}
