% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnct_create.R
\name{create_functions}
\alias{create_functions}
\alias{create_base_data}
\alias{create_filtered_data}
\title{Create functions}
\usage{
create_base_data(con, input_source)

create_filtered_data(data, input)
}
\arguments{
\item{con}{a databae connection}

\item{input_source}{usually an object created by a sidebar
function. These objects tend to be \code{reactive()} outcomes from
\code{observe()} or \code{observeEvent()} calls within a module.}

\item{data}{a \code{reactive()} data frame like object usually
created from another \code{create_function}. Considering
raw data is gathered through a PostgresSQL connection
these reactive objects tend to be \code{tbl_lazy}.}

\item{input}{shiny input.}
}
\description{
These functions are all use other functions to create
\code{shiny::reactive()} objects usually used in modules.
Any function that uses a \code{reactive()} call is considered
to be creating an object that is then displayed in the UI.
}
\details{
\code{create_base_data()} creates base data to be used in interactive map.

\code{create_filtered_data()} creates filtered data to be used in interactive map.
}
