% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnct_start_db_con.R
\name{db_connections}
\alias{db_connections}
\alias{start_db_con}
\title{Start Database Connection}
\usage{
start_db_con(db_name = NULL)
}
\arguments{
\item{db_name}{The name of the database we are wanting to connect to
Default is \code{NULL} and will use \code{gbif.duckdb}}
}
\value{
Returns a connection object.
}
\description{
This function uses environment supplied variables, \code{{DBI}},
and \code{{duckdb}} to connect to a DuckDB database.
}
