% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnct_display.R
\name{display_functions}
\alias{display_functions}
\alias{display_slider}
\alias{display_map}
\title{Display functions}
\usage{
display_slider(data, output, output_id, session)

display_map(data, output, output_id)
}
\arguments{
\item{data}{a reactive data object for a given variable.
\code{observe()} or \code{observeEvent()} calls within a module.}

\item{output}{an output within a shiny module.}

\item{output_id}{the id name of the output, e.g., \code{"summary_histogram"}.}

\item{session}{module session.}
}
\description{
These functions are all display different things for the user.
They function by wrapping the given display object (e.g.,
slider) with a \verb{render*()} function from \code{shiny}.
}
