% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnct_get.R
\name{get_functions}
\alias{get_functions}
\alias{get_dropdown_options}
\title{Get functions}
\usage{
get_dropdown_options(
  con,
  tbl_name,
  value_col,
  filter_col = NULL,
  filter_val = NULL
)
}
\arguments{
\item{con}{a valid \code{DBI} connection to a DuckDB database.}

\item{tbl_name}{a given name for the a table of interest in the database.}

\item{value_col}{a given name for the column of interest e.g., \code{scientific_name}.}

\item{filter_col}{a given name of for the column used to filter e.g., \code{country}.}

\item{filter_val}{a given name for the object that will be filtered against.}
}
\description{
These functions are very important as they create SQL to
then get data from DuckDB database, whether that is
schemas, raw values or other information about the database.
Each function creates a SQL string that is excuted on a
given table in the database.
}
\details{
\code{get_dropdown_options()} gets the options and values for the dropdowns. It
functions based on which countries are selected.
}
