% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_home_sidebar.R
\name{home_sidebar_module}
\alias{home_sidebar_module}
\alias{home_sidebar_ui}
\title{Home Sidebar}
\usage{
home_sidebar_ui(id)
}
\arguments{
\item{id}{the shiny namespace id name (i.e., \code{"home_sidebar"}).}
}
\description{
Provides the sidebar for the home table pane.
}
\details{
\code{home_sidebar_ui()} provides the home sidebar user interface.
}
