% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnct_map.R
\name{map_functions}
\alias{map_functions}
\alias{base_map}
\title{Map functions}
\usage{
base_map(df)
}
\arguments{
\item{df}{a \code{reactive()} object containing the data to be
represented on the map.}
}
\description{
These functions create \code{{leaflet}} maps to display the data
}
\details{
If df is null or does not have any rows then
base map with message iss displayed
}
