% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnct_test_db_con.R
\name{test_functions}
\alias{test_functions}
\alias{make_test_con}
\title{Make Test DB Connection}
\usage{
make_test_con(tbl_name, df)
}
\arguments{
\item{tbl_name}{a \code{character} that is the name of the tbl to test}

\item{df}{a \code{data.frame} to write}
}
\description{
Make test db connection
}
