% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_poland_occ}
\alias{test_poland_occ}
\title{A \code{data.frame} containing the four species from poland used for testing}
\format{
\code{data.frame} containing 10 rows and 42 variables
}
\usage{
test_poland_occ
}
\description{
A \code{data.frame} containing the four species from poland used for testing
}
\keyword{datasets}
