% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_poland_spp}
\alias{test_poland_spp}
\title{A \code{data.frame} containing the names four species from poland}
\format{
\code{data.frame} containing 4 rows and 4 variables
\describe{
\item{country}{The country name)}
\item{scientific_name}{The scientific name}
\item{vernacular_name}{The vernacular name}
\item{display_label}{The combined of scientific and vernacular name}
}
}
\usage{
test_poland_spp
}
\description{
A \code{data.frame} containing the names four species from poland
}
\keyword{datasets}
