% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_glatar.R
\name{gbif_app}
\alias{gbif_app}
\title{gbif Shiny App}
\usage{
gbif_app()
}
\description{
This function which has no arguments starts up the app,
creates the user interface (UI) and server and combines them
to make the app.
}
