% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnct_get.R
\name{get_functions}
\alias{get_functions}
\alias{get_map_data}
\alias{get_tbl}
\title{Get functions}
\usage{
get_map_data(con, input)

get_tbl(con, tbl_name, select = NULL)
}
\arguments{
\item{con}{a database connection.}

\item{input}{a input object produced by a sidebar.}

\item{tbl_name}{a given name for the a table of interest in the database.}

\item{select}{a \code{vector} of columns to select.}
}
\description{
These functions are very important as they create SQL to
then get data from DuckDB database, whether that is
schemas, raw values or other information about the database.
Each function creates a SQL string that is excuted on a
given table in the database.
}
\details{
\code{get_map_data()} selects and returns map data to be ploted on the map

\code{get_tbl()} selects and returns a query for the given table
}
