% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnct_refresh.R
\name{refresh_functions}
\alias{refresh_functions}
\alias{refresh_species_by_country}
\alias{refresh_selection}
\title{Refresh sidebar filter choices based on current selections}
\usage{
refresh_species_by_country(input, session, con, updating)

refresh_selection(session, name, options)
}
\arguments{
\item{input}{the module's \code{input} object}

\item{session}{the module's \code{session} object}

\item{con}{a DBI/duckdb connection}

\item{updating}{a \code{reactiveVal()} used as a lock to prevent circular updates}
}
\description{
Refresh sidebar filter choices based on current selections
}
\details{
\code{refresh_species_by_country()} creates updated filters based on changes in the
country selected.
}
