% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnct_search.R
\name{search_functions}
\alias{search_functions}
\alias{search_species}
\title{Search functions}
\usage{
search_species(session, input, con, selected_country = NULL)
}
\arguments{
\item{session}{shiny session}

\item{input}{shiny input}

\item{con}{a databaese connection}
}
\value{
returns a filtered data object that can then be displayed.
}
\description{
Search functions allow for a dynamic search bar to be displayed to search
taxa
}
