% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_poland_mm}
\alias{test_poland_mm}
\title{A \code{data.frame} containing the four species from poland used for testing}
\format{
\code{data.frame} containing 4 rows and 9 variables
}
\usage{
test_poland_mm
}
\description{
A \code{data.frame} containing the four species from poland used for testing
}
\keyword{datasets}
